+++ matlab/+cepton_sdk/+c/get_version_string.m
@@ -1,0 +1,3 @@
+function result = get_version_string()
+    result = cepton_sdk.c.call('cepton_sdk_get_version_string');
+end