+++ matlab/+cepton_sdk/+capture_replay/get_filename.m
@@ -1,0 +1,3 @@
+function result = get_filename()
+    result = cepton_sdk.c.call('cepton_sdk_capture_replay_get_filename');
+end