+++ python/cepton_util/__init__.py
@@ -1,10 +1,10 @@
 __author__ = "Cepton Technologies"
-__version__ = "1.9.3"
+__version__ = "1.10.1"
 
 import cepton_util.common
 
 _all_builder = cepton_util.common.AllBuilder(__name__)
 
-from cepton_util.common import *  # isort:skip
+from cepton_util.common import *  # noqa isort:skip
 
 __all__ = _all_builder.get()